refactor SimpleDocumentUploader
